Handling

+++ src/models.py
@@ -1,19 +1,19 @@
 from src import db
 from sqlalchemy.sql import func
-from sqlalchemy.dialects.mysql import TINYINT
+from sqlalchemy.dialects.postgresql import SMALLINT
 
 class User(db.Model):
     __tablename__ = 'user'
 
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     age = db.Column(db.String(255), nullable=False)
-    salary = db.Column(TINYINT(), nullable=False)
-    job = db.Column(TINYINT(), nullable=False)
+    salary = db.Column(SMALLINT(), nullable=False)
+    job = db.Column(SMALLINT(), nullable=False)
     created_at = db.Column(db.TIMESTAMP, server_default=func.current_timestamp(), nullable=False)
     deleted_at = db.Column(db.TIMESTAMP, nullable=True)
     province = db.Column(db.String(255), nullable=False)
     updated_at = db.Column(db.TIMESTAMP, server_default=func.current_timestamp(), onupdate=func.current_timestamp(), nullable=True)
-    gender = db.Column(TINYINT(), nullable=False)
+    gender = db.Column(SMALLINT(), nullable=False)
 
     def __repr__(self):
         return f'<User {self.id}, Age: {self.age}, Gender: {self.gender}>'
